requirements.txt
django
django-dotenv
djangorestframework
django-cors-headers
django-rest-knox
PyYAML
cloudinary
requests
uplink
markdown
cloudinary
channels==3.0.5
channels-radis
psycopg2-binary
